Undo Test

+++ Bubble/src/Bubble/Debug/Intrumentor.h
@@ -27,7 +27,7 @@ namespace Bubble {
     {
         std::string Name;
         long long Start, End;
-        size_t ThreadID;
+        uint32_t ThreadID;
     };
 
     struct InstrumentationSession
@@ -125,7 +125,7 @@ namespace Bubble {
             long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
             long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-            size_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+            uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
             Instrumentor::Get().WriteProfile({ m_Name, start, end, threadID });
 
             m_Stopped = true;
